requirements.txt
amqp==5.0.6
billiard==3.6.4.0
black==21.12b0
celery==5.2.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
flake8==4.0.1
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.2.2
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
pause==0.3
platformdirs==2.4.0
prompt-toolkit==3.0.23
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
tomli==1.2.2
typing-extensions==4.0.1
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
